Add ingest throttle by scheduling.maxTxnsPerSec

+++ hedera-node/hedera-config/src/main/java/com/hedera/node/config/data/SchedulingConfig.java
@@ -26,12 +26,9 @@ import com.swirlds.config.api.ConfigProperty;
 // spotless:off
 @ConfigData("scheduling")
 public record SchedulingConfig(
-        /**
-         * The fraction of total network capacity that can be scheduled to execute in a second.
-         */
         @ConfigProperty(defaultValue = "1:10") ScaleFactor schedulableCapacityFraction,
         @ConfigProperty(defaultValue = "false") @NetworkProperty boolean longTermEnabled,
-        @ConfigProperty(defaultValue = "100") @NetworkProperty long maxTxnPerSec,
+        @ConfigProperty(defaultValue = "100") @NetworkProperty int maxTxnPerSec,
         @ConfigProperty(defaultValue = "10000000") @NetworkProperty long maxNumber,
         @ConfigProperty(defaultValue = "5356800") @NetworkProperty long maxExpirationFutureSeconds,
         @ConfigProperty(defaultValue =
